Update repo_create_commit callers to pass git_extra_headers
Summary: `repo_create_commit` and the underlying logic for creating the changeset present in `create_changeset` are called from both EdenAPI and SCS. This diff passes the `git_extra_headers` field from all the callsites to this method. This field will only be populated in `git-import` and `git-remote-import`, which would be done in the next diff.

Differential Revision: D43911730

fbshipit-source-id: ff784460687f58b276f83e6b0b7b21a6ee962e39