Expose git_extra_headers via commit_info and history methods
Summary: The changes made in the previous diff allow to expose the git_extra_headers through `commit_info` and `history` methods so that the information can be read via clients. I also updated SCS to make use of `git_extra_headers` and to print them, in case they are UTF-8 encoded.

Reviewed By: mitrandir77

Differential Revision: D43876601

fbshipit-source-id: cb8f7836acfa2f66a2ace20d5dc9d8d1aaecf394