enable grep tests on mac
Summary:
hg grep under the hood is `hg st | xargs grep`. The return code is the xargs
return code. xargs on mac and linux behave a little differently.

according the the linux man page:
```
EXIT STATUS
       xargs exits with the following status:
       0 if it succeeds
       123 if any invocation of the command exited with status 1-125
```

according to the mac man page:
```
EXIT STATUS
     The xargs utility exits with a value of 0 if no error occurs.  If utility cannot be found, xargs
     exits with a value of 127, otherwise if utility cannot be executed, xargs exits with a value of
     126.  If any other error occurs, xargs exits with a value of 1.
```

So a failure on linux in grep will result in a 123 code while on mac it will
result in a 1 exit code. I wish mac behaved like linux, but it doesn't. So lets
teach the tests to accept the xargs return code based on the platform.

Reviewed By: mshroyer

Differential Revision: D43924480

fbshipit-source-id: 0e6c2ec2c6a7553a1aed7c072f0464e5d1fa64c7