Delete .rlsconfig files
Summary:
The last mention of "rlsconfig" I can find on Workplace is from 3.5 years ago in 2019.

RLS has been deprecated officially since https://blog.rust-lang.org/2022/07/01/RLS-deprecation.html and unofficially for multiple years before that.

Of the 14 nonempty rlsconfig files, not a single one had been updated to platform010, strongly suggesting they are not being used. Of the 14, more than half had not even been updated from platform007 to platform009. The ones that were ever updated to platform009 have a test plan like D28345242 which does not inspire confidence that the person knows what the file is for.

Reviewed By: zertosh, edward-shen

Differential Revision: D43964314

fbshipit-source-id: 18f43db50dd9ebda0d4846cd785e2fe34a938eea