Implement TryAssemblyLine trait
Summary:
As I complain about [here](https://www.internalfb.com/diff/D43395160?dst_version_fbid=514424547472616&transaction_fbid=777341240631879), creating a trait for `TryAssemblyLine` is very tricky because we'd ideally like to name some simple closures but can't.

But it turns out it **is** possible. I needed to create a struct and implement `FnMut` for it, as well as naming some other types I really don't care about (like `future::Either<F::Output, future::Ready<Result<T, Err>>>`), which is not very ergonomic in Rust. And the struct is pretty complicated. It does work, though, and does not need copypasting.

This struct is generic enough to be used in the future if we add more functions to `AssemblyLine`.

Reviewed By: mitrandir77

Differential Revision: D43503770

fbshipit-source-id: 1d082d5f714b62d6d62b317f5e2b117de95f7088