Make create_bookmark category-aware
Summary:
This allows user of this API to create bookmarks that can represent git tags or git notes in addition to the default bookmarks that map to git branches.

Note that annotated tags are not covered in this diff. What differentiates a lightweight tag from an annotated tag is the changeset it points at. Creating the changeset is independent from creating the bookmark.

Differential Revision: D43775947

fbshipit-source-id: 6865c091a29917528fccfdca6db1af26a6f1fcb4