Add category column
Summary:
This was intentionally omitted when the `BookmarkCategory` was introduced in D43444397.

The reasoning behind that omission was that this is used during `move_bookmark` and tags are not expected to move.

While tag moves are more rare than branch moves, they can still happen:
git provides a mechanism for moving them (`git tag -f`) and it can also occur indirectly if a tag is removed and re-added somewhere else in git, followed by a push, where live git sync will need to move the bookmark.

Don't special-case the category in the bookmark_update_log: tags and notes should behave like any other bookmarks.

Note: D43908929 is the configerator counterpart to this change and should land and rollout first.

Differential Revision: D43908709

fbshipit-source-id: 4423dc6dc9a5e7f2ffe9d4e8f62741f7ee38a831