Make resolve_bookmark category-aware
Summary:
This allows us to refer explicitely to a specific bookmark whether it's a tag, a note or a branch.

Prior to this change, we could only refer to bookmarks of type `Branch` (which is the default value and the case for any bookmark that was not imported from a git repository).

Differential Revision: D43909897

fbshipit-source-id: f767a3bcf236652aae48b63a4724355187282a46