Fix clippy warnings (2/2)
Summary:
This fixes the remaining clippy errors.

The only thing l haven't fixed is that clippy crashes when trying to parse `eden/mononoke/blobstore:multiplexedblob_wal`. This doesn't generate warnings and the command still succeeds, but it's annoying. I haven't been able to track down why, so I'm leaving it be.

Reviewed By: YousefSalama

Differential Revision: D43980977

fbshipit-source-id: 28ddcce5894903d6e0f980cd04b586dc1afc00c6