workingcopy: pull logic out of FileChangeDetector
Summary: I want to share this logic w/ a different file detector architecture, so pull it out and consolidate into a single free function. This has one concrete advantage of avoiding a duplicate treestate lookup for every file. This may not be the final form, but it's easiest for me right now.

Reviewed By: quark-zju

Differential Revision: D43626600

fbshipit-source-id: 6ee4355110fc76f2542bf78d8be348e894e73cbd