workingcopy: add more matcher tests for watchman status
Summary:
Add tests to make sure we mark files NEED_CHECK properly even if the given matcher doesn't match the files. Although we try to limit work based on the matcher, we need to at least record NEED_CHECK for files reported by watchman since we will update the watchman clock in the treestate (and won't get notified about the non-matching file next time).

Note that this isn't right for ignored files which should not be tracked in the treestate by default. Will address in following diffs.

Reviewed By: quark-zju

Differential Revision: D43760941

fbshipit-source-id: f445f9576323b730e72d9df1acce2a26b8a3ef50