workingcopy: make FileChangeDetectorTrait parallel friendly
Summary:
I want to parallelize the work this object does, but the trait interface is serial (and synchronous):

```
pub trait FileChangeDetectorTrait {
    // Check if file size and mtime are unchanged (otherwise queue "maybe changed" result).
    fn has_changed(&mut self, ts: &mut TreeState, path: &RepoPath) -> Result<FileChangeResult>;

    // Resolve "maybe changed" files by checking contents.
    fn resolve_maybes(&self) -> Box<dyn Iterator<Item = Result<ResolvedFileChangeResult>> + Send>;
}
```

It also was a bit annoying to use because it reported results w/ two different types, so the client needed duplicate handling. The new trait allows for asynchronous processing and only returns a single kind of result:

```
// Converting to an iterator signals you are done submitting paths.
pub trait FileChangeDetectorTrait: IntoIterator<Item = Result<ResolvedFileChangeResult>> {
    fn submit(&mut self, ts: &mut TreeState, path: &RepoPath);
}
```

I converted the existing concrete implementation to the new trait. It requires the object to buffer more results, but not any file data, so it shouldn't have much of an impact memory wise.

Reviewed By: quark-zju

Differential Revision: D43626597

fbshipit-source-id: 4ffcd9b870ee8a4fba94aa3972cc4e1a299ebec5