Add endpoint to create a tag annotation
Summary:
Until now, we could use `BookmarkCategory` to create bookmarks referring to branches, tags or notes; but we didn't have a mechanism to create a tag annotation, which makes the difference between a lightweight tag and an annotated tag (notes should also always be annotated).

With this change, we have all the mechanics to create, delete or move any bookmark that can be imported from git.

Differential Revision: D43775948

fbshipit-source-id: 5241f94f190fcfa8c64747fd862cf70a3ed8909d