treestate: clean up metadata API a bit
Summary:
The main thing is to add methods to TreeState to get and set the entire deserialized metadata. The existing get_metadata_by_key would deserialize the metadata each time it was called.

I removed the Metadata type from most places. It is a wrapper type only used for serialization, so it probably shouldn't be public at all.

I also renamed set_metadata_by_keys to update_metadata, which goes better with the new set_metadata.

Reviewed By: zzl0

Differential Revision: D43970591

fbshipit-source-id: 203cc68acd6476e5e11608fb4858c9abb77ec03c