workingcopy: don't pass TreeState to FileChangeDetector
Summary: Now that normalized_get() is available on the TreeState, it's less important to let the FileChangeDetector query the treestate. Further, I want to add some logic based on the treestate entry to watchmanfs, and having the caller pass it in means the caller will have it in-hand to use without a redundant query.

Reviewed By: zzl0

Differential Revision: D43970587

fbshipit-source-id: 9b7e7cdeb5654f0fabd5854115af572c32ddd833