document the ways FSCK was broken and the ways we have fixed it.
Summary:
FSCK is complicated on Windows, and it's broken in a lot of.

We have fixes in the works, but let's get this knowledge out of my
brain and into a common place.

I originally wrote it to be a note, but Xavier said it was too long. So
I'm putting it in the docs instead.

happy reading :)

Reviewed By: mshroyer

Differential Revision: D43857890

fbshipit-source-id: 2534206106eb0d48297462ce796c9c17f6618c7a