Enable StartFakeEdenFSTest
Summary:
This is another case where we had issues with the long default temp path on
macOS.  Here, fake_edenfs was failing with SIGABRT on startup due to trying to
open a unix-domain socket with too long of a path.

D43925614 addressed this for tests based on EdenTestCase, but this doesn't help
with StartFakeEdenFSTestBase which only inherits from EdenTestCaseBase.  So
this diff moves the environment variable override one level up in the type
hierarchy to cover more test cases.

Reviewed By: kmancini

Differential Revision: D43988120

fbshipit-source-id: 9eca299c42aea819d3e93ed5bb408b4a4783d8ee