mononoke_api: extend create commit checks to support stacked changes
Summary:
The create commit method performs a number of checks to ensure that the requested commit is valid.  This is in addition to the bonsai changeset checks that ensure that a bonsai changeset is internally consistent.

Extend these methods so that they can operate when changes are stacked.  The checks use a path tree of stacked changes to validate that file operations earlier in the stack are consistent with file operations of each commit.

Reviewed By: mitrandir77

Differential Revision: D44032675

fbshipit-source-id: e245dd997d57dee1aebe76605c80bff9e3586be9