mononoke_api: allow creation of stacks of commits
Summary:
Add `create_changeset_stack` to `RepoContext` to allow creation of a stack of commits in one operation.

In addition to previous checks, we must build a stack of the accumulated changes for each of the earlier commits in the stack so that we can check that the stack is internally consistent.  For example, a file can only be deleted once in the stack, or it can be created and then deleted in the same stack.

Reviewed By: mitrandir77

Differential Revision: D44032676

fbshipit-source-id: f08a7df39c52c9de9d6efc728321112ac5ee60f8