refactor info from fetched commits and files
Summary: We want to include a fetch start and end timestamp in our fetches from sl log and sl status. This will let us clear state for sufficiently new fetches after operations run. To support this, this diff refactors the types for how we report the fetched data. Essentially, we wrap the actual files/commits into a containing object. By making this a Result<T>, we don't need to use the error part of the event emitter anymore.

Reviewed By: muirdm

Differential Revision: D44045268

fbshipit-source-id: dda9e765ca4230c2d4589870ec5276beae73e078