Force clear optimistic state after a fetch after an operation
Summary:
This diff adds a fallback mechanism to clear out "bad" optimistic state. This prevents cases where the UI would appear stuck loading.

Optimistic state appliers each define how that operation should be previewed, and importantly, when that preview is no longer necessary.

Currently, there are some cases where a preview applier will never see the sufficient condition for the optimistic state to resolve. This is largely caused by commit successions.

For an example:
- rebase a stack of commits A,B,C
- while it's rebasing, click "goto B"
- our optimistic state shows "youre moving here" on commit B, by looking for the commit in the tree with B's hash
- after the rebase completes, B has been succeeded by a new version, B2
- eventually, we get a new list of commits from `sl log`, including the new commit B2 with `isHead: true`.
- the optimistic state for the "goto B" operation, however, is waiting for you to checkout commit B, not B2.
- so the "you were here..." / "you're moving here..." are stuck forever, since you'll never checkout the now-obsolete B hash again.

Note that solutions that alter the optimistic state applier for GotoOperations are tricky, for example looking for any time the head commit changes doesn't work in the face of multiple checkouts, and trying to associate commits by something other than their commit hash is also tricky and error-prone.

The proper way to fix this is to watch for commit succession events, and update the optimistic state applier to look for something like `successor(commitHash)`. This actually basically matches what sl does.

This diff takes a different angle and provides a rougher fallback to clear out all optimistic state. In the repro steps above, once we got the new list of commits from `sl log`, if that fetch was started and finished AFTER the goto operation, we know that that `sl log` MUST account for all changes from the goto operation. So upon seeing ANY successful fetch after the goto, we can clear out its optimistic state.

This is not quite as nice of an experience, because some optimistic state cases may not be handled and thus cause some UI jumping around. But it guarantees we will not be stuck with "bad" optimistic state any longer than 1 `sl log` command after the operation finishes.

Even once we add a succession system, this clearing of optimistic state will still be useful.

Reviewed By: quark-zju

Differential Revision: D44045265

fbshipit-source-id: 047340cebbecef05355718cf06f03a68039b7b09