status: fix rust flags comparison
Summary:
Fix rust status when dirstate does not contain executable/symlink flags for a tracked file, and the file's flags have diverged from the manifest.

To achieve this we pipe the filesystem metadata long to the content check since the content check already looks up all the files in the manifest. Before actually doing the content check, it compares the on-disk flags to the manifest flags.

Reviewed By: quark-zju

Differential Revision: D43970583

fbshipit-source-id: 544acfa355cf9d0d2acf59ad9143de6dac349235