sparse: kill marker file detection
Summary: Sparse profiles supported configuring a list of "unsafe" marker files that would cause an error if they accidentally were included in a sparse profile (presumably to avoid unneeded sparse profile growth). However, this feature is not in use anymore, and I don't want to implement it in Rust, so kill it.

Reviewed By: quark-zju

Differential Revision: D43970582

fbshipit-source-id: a6576b8004cc41b3e6f2f0e0040b54e54af2bdc4