inode: enhance checkout telemetry
Summary:
Understanding how many trees and blobs were accessed but not fetched during
checkout is a useful metric that informs us as to how much was served from
caches. This is useful as checkout is more or less proportional in time to the
number of accessed inodes and the number of accessed trees+blobs will thus tell
us how many TreeInode and FileInode were loaded and traversed.

Reviewed By: chadaustin

Differential Revision: D44117153

fbshipit-source-id: 56991c51f2e4f501486d64ab5598f149f2708b77