Standalone platform for Tauri app
Summary: Add a new platform entry point for the "standalone" Tauri app (see next diff). This will let us spawn `sl web` with `--platform standalone`, in order to get specific implementations of features for the tauri app. We actually mostly just need to use browser platform features for this. But at the very least we want to be able to have the "platform.name" be "standalone" to differentiate it from the browser.

Reviewed By: quark-zju

Differential Revision: D43888460

fbshipit-source-id: ff0515c35c2d02ae9d69117a2f4a947a531f7934