Standalone Tauri app
Summary:
See this document for more information:
https://docs.google.com/document/d/1kSdc86XNyPMt_rTB8uggYOZQ0hf4k7Fq4NhK1N3taZE/edit#

The idea here is to make a standalone application to access ISL, rather than using `sl web` to open ISL in a browser. This makes ISL feel more native and lets it use OS windowing features.

Since ISL is a web application, we need a browser-like technology like electron. Electron is quite bloated, so Tauri is a good alternative for our use case.

We only need to run `sl web` and embed that into a webview with Tauri. We don't need to produce entirely new javascript bundles for Tauri. This means the Tauri application shell can be updated independently of the actual ISL client/server code.

I don't intend to pursue this much futher right now, since the next steps involve figuring out distribution mechanisms and build pipelines. But I figured it was worth publishing and even landing this so it's easily accessible in the near future.

Reviewed By: quark-zju

Differential Revision: D40149722

fbshipit-source-id: 8225ffd54d77141254166676af65b991d4a1a3c1