Fix misaligned blue highlight bar
Summary:
UX nit, the blue line under the "You are here" did not go all the way to the commit bubble when on a draft commit, which looked incorrect.

Another small nit is that the "You are here" branching path when on a public commit did not fully line up with the bubble either, which left a very small 1-2 px hard border on the path.

Reviewed By: quark-zju

Differential Revision: D44141061

fbshipit-source-id: be79870ea6688074ef36b190862297c385391c19