Unify subscriptions from client and handle unsubscribing
Summary:
We had a very loose "subscription" system between the client and server. The client would send a message like `subscribeUncommittedChanges`, then the server would set up an `subscribeToUncommittedChanges` to send messages back. But there was no way to stop listening, and if we sent multiple subscriptions they would just get duplicated. We had a hacky deduplication in the previous diff.

instead, we should use a real `subscriptionID` to identify which responses are for the right subscriptions. This way, the client can also send `unsubscribe` messages.

Refactoring a bit into dedicated `subscribe` / `unsubscribe` / `subscriptionResult` messages makes it a bit more clear and less duplicated.

The client also refactors the recoil `atom` effects to a unified `subscriptionEffects` utility.

This is a rather pervasive change, so it broke a lot of tests. It's mostly just because we rename the subscription events coming from the server, which are used in basically all the tests.

Reviewed By: quark-zju

Differential Revision: D44102868

fbshipit-source-id: 4c0f5347d6a5ea57f58c6db005178ddce2310cfc