Support showing ISL in the sidebar
Summary:
In VS Code, we embed ISL as a webview. The default for webviews is to appear as an panel like an editor, as a tab in the main view. This works, but can feel somewhat heavy-weight. There's no icon on the sidebar, so you have to use a command or keybind. It will remain as a tab, which you can either find or run the command again. The tab takes up a spot that could otherwise be another editor. This is sort of the vscode-sanctioned way to do it.

But VS Code also allows you to put webviews in the sidebar as a—horribly named—"webview view". It works very similarly in the code, we just slightly change how the webview is created. This is nice because it adds a sidebar item (with the sapling logo), which you can click to open ISL. You can actually drag this to the right side or the bottom panel, too, which makes it much more customizable than editors are.

## Size
In terms of UX, it's similar, though having it in the sidebar makes it more likely to be used with a low width. Most other sidebars only need 300-400px of width. ISL benefits from having 1000px+, so that you can have the commit tree, with long-ish titles, plus the commit info sidebar, itself with decent width to render descriptions nicely.

But by expanding the ISL sidebar, you also expand the other sidebars, like the explorer, which is not what you want. Thus, I expect users will be likely to leave ISL at low screen widths. **Thus, we will need to make ISL behave better at low widths**. For example, we could make the commit info sidebar into a tab at the top instead of a sidebar. Having a sidebar in a sidebar is confusing anyway.

There is a cool workaround to the resizing problem, which some power users may find (or maybe we can NUX it). You can drag the ISL sidebar to the right "secondary" sidebar instead of the left one. This is another expandable sidebar, however it stores its width separately, so you could make it have a larger width and only have ISL in it. We [cannot configure it to appear on the right by default](https://code.visualstudio.com/api/ux-guidelines/sidebars#secondary-sidebar), which is why this would need some kind of nux.

This diff has a small helper for behaving well at small sizes, we default the commit info view sidebar to be collapsed if the screen width is smaller than 500px.

## Configuration

We can allow either option for now, while we figure out which one is better. Maybe we'll always support both since it's not that complicated.

I added a vscode configuration for whether we use ISL as a panel (editor-like) or a view (sidebar). We don't want to support both simultaneously—it gets confusing and we generally expect only one ISL at a time.

VS Code seems to have a small bug where if you don't start vscode with the showInSidebar setting enabled, it won't ever show the new viewContainer. Restarting makes it work, so I just say "restart to take effect" in the setting description.

Reviewed By: quark-zju

Differential Revision: D43806332

fbshipit-source-id: f54e7235982c6d7461c06684169f33e57ee320ab