Don't ignore subscriptions after the first
Summary:
In the previous diff, we allow ISL to live as a sidebar in VS Code. When the sidebar is hidden and re-opened, our client re-connects to the server, and we need to resubscribe to subscriptions for uncommitted changes and commits.

We had some code in here to dedupe subscriptions. This gets in the way of us properly resubscribing.

**Why were we deduping?**
 In development mode, react runs effects twice, disposing them in between invocations. This helps catch bugs where you don't handle unsubscribing properly.

Well, we weren't handling unsubscribing properly. The terrible hacky fix is to dedupe on the server. The better fix is to have the client send an "unsubscribe" message, and just live with the slight overhead of double calls in development mode.

**What does this diff do?**
This diff starts the process by removing our server-side deduping. This also fixes the vscode sidebar behavior so it actually works. This is the easy part.

Reviewed By: quark-zju

Differential Revision: D43806331

fbshipit-source-id: dd86693297d679f7d935cbf37dddf56e24477774