don't drop MononokeApp in wait_until_terminated before termination signal to avoid dropping ConfigHandle
Summary:
When MononokeApp is dropped, the ConfigHandle inside of it gets dropped, and the config updater thread stops receiving updates causing config hot reloading to stop working. This was happening in scs, the land service, and the derived data service (not sure if this is also happening in Mononoke server).
This diff adds an explicit drop in the server closure in wait_until_terminated to make sure MononokeApp is not dropped before receiving a termination signal.

Reviewed By: Croohand

Differential Revision: D44206206

fbshipit-source-id: 7322f516c0dcbf4668cba811bdb582f0e5e2f11c