Stack actions
Summary:
Add a new notion of "stack actions" which appear as buttons at the bottom of a stack of commits. I would love feedback on the concept & look, hence RFC.

These actions will let you do things that affect the entire stack. For now, this means submitting the entire stack for code review, or re-submitting if commits have already been submitted.

In the future, this would be an ideal place to include other features like "rebase stack onto main" or "edit stack" (for histedit) or more advanced submit actions like "resubmit only changed diffs" (diff with phabricator).

In the old ISL 1.0, these kinds of features had to be awkwardly put next to the bottom commit in a stack. I think that's inaccurate—we want these actions to be associated with the entire stack.

The UI change here is to add stack actions below the stack. If your stack of commits has no available actions, nothing is rendered. If you have only one commit, it's not a stack and therefore doesn't show stack actions (you can do normal actions to submit/rebase/etc).

We may find the list of actions will grow, so in the next diff I add a concept of "more actions", which are de-emphasized actions you can take which appear as a [...] menu you click to see like a context menu.

In this diff, I focus on the re-submit stack feature. I leave it up to each code review provider to describe what submitting features a given stack supports. This allows the review provider to make decisions based on code review status. For example, internally we don't allow re-submitting Closed diffs.

The downside of adding this feature is that we're elongating all our stacks vertically by one row. If you have lots of stacks, this would reduce information density. My assumption is that people rarely work with more than 1-3 stacks, and they are usually of moderate height 2-8 commits, such that this extra height is not oppressive, while being in a useful and obvious location.

I use the icon-button style to make the stack actions not overly noticeable. This is consistent with icon-button usage elsewhere in the UI.

Reviewed By: quark-zju

Differential Revision: D43547204

fbshipit-source-id: 541574d755dd845ca6145e3a09f29b70b7cc723f