Highlight commits to submit when hovering
Summary:
Hovering on the "submit stack" button should hint at what's going to be submitted, since it may not be obvious. This highlight can be a simple shadow/border on the commit's bubbles

We can add a new component to make this declarative.

One small edge case to consider is if the component being hovered unmounts. This means `onMouseLeave` won't trigger, leaving the hover stuck. We just need to have an effect to sync the highlight state back to empty on unmount.

Reviewed By: zzl0

Differential Revision: D44110879

fbshipit-source-id: ded52fe7c721419abdb9d478d983985895ffe1e8