Allow submitting specific diffs for review
Summary:
Add an argument to the codeReviewProvider-specific submitOperation for a set of commits. This lets you express submit operations for individual commits or stacks of multiple commits, or generate the appropriate submit operation which would work for multiple commits.

This allows us to implement a submit-selected commits feature. Sadly, this only works internally right now, as `sl pr submit` and `sl ghstack submit` only work for the current head commit. Maybe this will change in the future.

Of course, we can only really make use of this parameter to the submitOperation if the code review provider supports it. In the next diff we'll add an API that lets a codeReviewProvider express whether it is able to submit a particular diff. This means we don't intend to call submitOperation with arguments that a particular code review provider can't handle. We merely leave the option available to providers.

Reviewed By: quark-zju

Differential Revision: D43547201

fbshipit-source-id: ae9e5678a42586cd764093cefa81dcf344de7d3c