gotham_ext: add stream stats gathering
Summary:
A gotham request consists of two parts.  The first part is a future, in which the handler resolves the request and produces the response headers.  The second part is the body which may optionally be a stream.  We added collection of future stats for the first part, but the majority of work may actually be done when computing the stream for the body.

In this diff we add stats collection for that body stream, if it is used.  Since this is separate from the future stats we already collected, we prefix the scuba columns with "stream_" to make it clear this is the stream.

Reviewed By: liubov-dmitrieva

Differential Revision: D44260233

fbshipit-source-id: 9820d3f30fb147740a95d8afc2fc1bf1fcc16aab