Publish liveness metric for ConfigUpdater process
Summary: Youssef accidentally discovered that config hot reloading was not working for SCS server when working on an oncall task. He then went ahead and fixed that issue in D44206206. However, we did not get any notification or alert about this problem and it would have stayed hidden much longer had we not accidentally detected it. This diff adds a liveness metric for the `ConfigUpdater` which will keep publishing every 5 mins as long as the `MononokeConfigs` object is live.

Reviewed By: YousefSalama

Differential Revision: D44293235

fbshipit-source-id: 6f54831c8b1f2154a23c021a7d2d8f01615a7d70