flush bookmark cache in the pull test
Summary:
Sometimes the tests run so fast that we're being served stale version of the
bookmarks.

This may cause the test to occasionally take the fast path where it expects to take the slow path, and has caused this test to be flaky since its creation (although not always flaky enough to be marked as such).

Differential Revision: D44298928

fbshipit-source-id: 046ce5bc4e005f3d443794202352bfdc5c80895a