Hide tooltip on mouse leave for simple tooltips
Summary:
An annoying behavior I noticed is that ISL tooltips would keep showing if your mouse entered the tooltip, even though it would leave the tooltip creator to get there.

For example, if you drag your mouse upwards and it passed a diff badge, then entered the next commit's title, and you tried to click to select that commit, it would still be inside the diff badge tooltip, so you'd click and it would open the diff in a new tab instead. This is especially annoying because the diff badge tooltip is shown on a delay, but the click targets are active immediately.

We can avoid this behavior by changing the `pointerEvents` for the tooltip, so that it doesn't register your cursor mouse enter at all.

However, we don't want to do this for complicated tooltips with custom rendering that aren't triggered via hover, since that would mean you couldn't click inside of them. For example, the settings tooltip. You need to be able to click inside that tooltip to change settings

Reviewed By: quark-zju

Differential Revision: D44261575

fbshipit-source-id: 9bcdbe4abc80bee0a1583da2c36bf6b5ee5955ed