Hide actions and uncommitted changes in multi-select sidebar
Summary:
When multiple commits are selected, we show them in the sidebar to add bulk actions like submitting all selected commits.

When rendering commits like this, we don't want to include the "You are here" uncommitted changes area, or any buttons like goto and uncommit. Those actions just don't belong in this context.

Reviewed By: quark-zju

Differential Revision: D44235554

fbshipit-source-id: 1f80b78350f9457f5e66139599f57c61c46eddc7