Persist draft checkbox via persistAtomToConfigEffect
Summary:
If you check the "submit as draft" checkbox, we should remember your selection across restarts of the app.

To do this, we want to use `sl config` to get and set a config value. This way, we're using the same config system that the rest of sl uses, which is both easy to do and consistent.

I wrote a helper that wraps this sl config persistence for any recoil atom via a recoil effect. This will make it very reusable on any other bits of UI that we want to persist (e.g. changing how uncommitted changes are rendered or whatever)

Reviewed By: quark-zju

Differential Revision: D44278888

fbshipit-source-id: fe024fb3518e541212eb41b29c2c3007a43012ef