Use succeedable revset when submitting specific commits
Summary:
Thanks to optimistic state, it's possible to queue up a submit for a specific commit. But using the hash would lead to trying to submit the obsolete commit when it gets succeeded by any operations before it in the queue.

We have a whole system for this, where we pass revset args in a special way to re-interpret them with a revset that skips past successors.

We've talked about changing this system a bit to track things more explicitly rather than just using `max(successors(${revset}))`, but for now this works ok, but we do need to be mindful to always pass revsets in this special way (that would be true even if we re-wrote this system to track successors manually)

This only applies for our internal submit flow, since ghstack and pr submit don't let you submit arbitrary commit hashes (you have to goto them first).

Also, for simplicity, we can remove the commit hash argument from submit when it's the head commit. This makes the command args simpler which is nice, and removes any issues around successions.

Reviewed By: quark-zju

Differential Revision: D44275963

fbshipit-source-id: 337837b1ef580711d2ba0495d21447a0678770d0