Add submit button to non-head selected commit
Summary:
If you select a commit, you see it in the sidebar. Depending on your code review provider, you may be able to submit this commit without `goto`-ing there.

We should add the `submit` button to the sidebar for these cases. But we defer to the review provider whether it supports submitting individual commits. If not, we don't show the button.

This also allows review providers to refuse to submit certain commits, such as already landed diffs.

Reviewed By: quark-zju

Differential Revision: D44238646

fbshipit-source-id: 2ca508a90539a7465d967bda8ad702723ef5a1e4