Submit as draft checkbox
Summary:
Add a draft checkbox next to the submit button in the commit info view, which controls whether the diff will be submit as a draft or not.

This makes sense to live next to the submit button in the commit info, since commit & submit / amend & submit are probably the most common ways to submit.

We'll use an interstitial for the "submit stack" functionality, since there's not real-estate for the draft checkbox under every stack.

The draft functionality actually varies between internal and github review providers. You can only create draft PRs at creation time, but not at update (resubmit) time. So we defer to the codereviewprovider to know if we can resubmit as a draft or not.

Reviewed By: quark-zju

Differential Revision: D44275874

fbshipit-source-id: 8b3be677ccce8d1299227f525301428987f546b9