Don't show file diff button next to merge conflicts
Summary: We added a diff button next to files when using vscode, so you can open a vscode diff view for a file. This doesn't make as much sense for merge conflicts, because the diff is not correct and the button makes it harder to find and click on the resolve conflict check button, which should be the most used one.

Reviewed By: quark-zju

Differential Revision: D44308276

fbshipit-source-id: 0663690774eb91cfd6896c83faecc6bf8a82b064