Move date revset inside smartlog()
Summary: I noticed that fetches for old commits were still sometimes failing, and I found that our revset was `smartlog() and date(-14) + .`, but we actually want `smartlog(date(-14) + .)` to calculate the smartlog from the set of recent heads and always include `.`. `smartlog(x)` is the better choice for us, since it automatically includes the ancestors of `x` in order to render the tree up to that point, which is exactly what we need.

Reviewed By: quark-zju

Differential Revision: D44276722

fbshipit-source-id: 6503471fb821dbc30d142ffbcbafb988652450cb