contrib: add a script to measure gca performance in hg and git
Summary:
I would like to refer to some performance data of `gca` in the Segmented
Changelog doc.

Initially I was trying to make `drawdag` fast since it can use an elegant
syntax `A001..A999` and can work for both git and hg formats
(`debugexportrevlog` can be used to create repos that the upstream Mercurial
can understand). However it turns out to be a lot of work to make `drawdag`
"just" fast so I ended up writing dedicated scripts to create repos with
millions of commits.

Reviewed By: muirdm

Differential Revision: D43373717

fbshipit-source-id: 5ea1753027cb7487099ed00007e4cc0beae8a09a