Add integration test runner
Summary:
Our tests right now are all unit tests, where the client has to mock responses from the server and the server can't tell if it's sending the messages the client actually understands.

Unit tests are great for mocking out specific scenarios, but we also want at least a handful of tests to validate that ISL can run on a real repository with real `sl` commands. This could catch a number of app-breaking bugs.

This diff adds the jest config / command to run integration tests, and a very simple first integration test that checks that commits appear in the UI and uncommitted changed files appear in the UI.

The idea is that we already run our UI tests in a node process with react-testing-library, which lets us fake render to a JSON "DOM" and then validate our UI DOM and click buttons and stuff. In that same node process, we can run `onClientConnection` to start and ISL server. We just need to hook the client code's messageBus up to our manually invoked `onClientConnection`, so that the client sends messages to the server. Little do they know it's all in the same process.

This is a very simple way to get integration tests to run, while also giving us all our normal react-testing-library UI assertions:
```
// inspect what was rendered
expect(screen.getByText('My Commit')).toBeInTheDocument();
// click buttons
fireEvent.click(screen.getByTestId('my-button'));
```

We can probably get fancier with our test repo, but right now it just runs `sl init --git .` to create an empty repo and adds a single file and commits it.

This all happens in a new repo per test run, so we should be able to run multiple integration tests in parallel. Multiple tests within one file would run in series.

Generally, I expect we'll still run most of our tests as unit tests, and use integration tests for specific complex workflows / scenarios that we want to make sure stay working, for example:
- basic commit rendering
- basic uncommitted changes
- writing commit messages
- handling merge conflicts
- interactive split / edit stack

Other things:

- Unfortunately, coverage doesn't seem to work outside the `<rootDir>` of the jest config, so we don't get coverage for the `isl-server` dir. Although we can get coverage information of the client.
- We could also do more end-to-end tests by spawning a standalone Tauri instance and doing tests with them and a real server. This would test the actual server<->client websocket code, but it may be harder to do assertions and more difficult to set up a test runner.

Reviewed By: muirdm, quark-zju

Differential Revision: D44184111

fbshipit-source-id: 164c36a6a7a741b3303236400d55b9a17cb8bede